+++ packages/layout/src/components/CollapsedIcon/style.ts
@@ -22,12 +22,12 @@ const genSiderMenuStyle: GenerateStyle<SiderMenuToken> = (token) => {
       alignItems: 'center',
       justifyContent: 'center',
       cursor: 'pointer',
-      color: token?.layout?.sider?.colorTextCollapsedButton,
-      backgroundColor: token?.layout?.sider?.colorBgCollapsedButton,
+      color: token.layout?.sider?.colorTextCollapsedButton,
+      backgroundColor: token.layout?.sider?.colorBgCollapsedButton,
       boxShadow:
         '0 2px 8px -2px rgba(0,0,0,0.05), 0 1px 4px -1px rgba(25,15,15,0.07), 0 0 1px 0 rgba(0,0,0,0.08)',
       '&:hover': {
-        color: token?.layout?.sider?.colorTextCollapsedButtonHover,
+        color: token.layout?.sider?.colorTextCollapsedButtonHover,
         boxShadow:
           '0 4px 16px -4px rgba(0,0,0,0.05), 0 2px 8px -2px rgba(25,15,15,0.07), 0 1px 2px 0 rgba(0,0,0,0.08)',
       },