+++ packages/layout/src/components/PageHeader/index.tsx
@@ -162,9 +162,8 @@ const renderChildren = (
 
 const PageHeader: React.FC<PageHeaderProps> = (props) => {
   const [compact, updateCompact] = React.useState<boolean>(false);
-  const onResize = ({ width }: { width: number }) => {
-    updateCompact(width < 768);
-  };
+
+  const onResize = ({ width }: { width: number }) => updateCompact(width < 768);
 
   const { getPrefixCls, direction } = React.useContext(
     ConfigProvider.ConfigContext,