+++ packages/layout/src/style/index.ts
@@ -27,10 +27,10 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
       width: '100%',
       height: '100%',
       [`${token.proComponentsCls}-base-menu`]: {
-        color: token?.layout?.sider?.colorTextMenu,
+        color: token.layout?.sider?.colorTextMenu,
         [`${token.antCls}-menu-sub`]: {
           backgroundColor: 'transparent!important',
-          color: token?.layout?.sider?.colorTextMenu,
+          color: token.layout?.sider?.colorTextMenu,
         },
         [`& ${token.antCls}-layout`]: {
           backgroundColor: 'transparent',
@@ -41,7 +41,7 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
             color: 'inherit',
           },
         [`&${token.antCls}-menu`]: {
-          color: token?.layout?.sider?.colorTextMenu,
+          color: token.layout?.sider?.colorTextMenu,
           [`${token.antCls}-menu-item`]: {
             '*': {
               transition: 'none !important',
@@ -69,50 +69,50 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
             ${token.antCls}-menu-item-active,
             ${token.antCls}-menu-submenu-active, 
             ${token.antCls}-menu-submenu-title:hover`]: {
-            color: token?.layout?.sider?.colorTextMenuActive,
+            color: token.layout?.sider?.colorTextMenuActive,
             borderRadius: token.borderRadius,
             [`${token.antCls}-menu-submenu-arrow`]: {
-              color: token?.layout?.sider?.colorTextMenuActive,
+              color: token.layout?.sider?.colorTextMenuActive,
             },
           },
         },
         [`&${token.antCls}-menu:not(${token.antCls}-menu-horizontal)`]: {
           [`${token.antCls}-menu-item-selected`]: {
-            backgroundColor: token?.layout?.sider?.colorBgMenuItemSelected,
+            backgroundColor: token.layout?.sider?.colorBgMenuItemSelected,
             borderRadius: token.borderRadius,
           },
           [`${token.antCls}-menu-item:hover, 
             ${token.antCls}-menu-item-active,
             ${token.antCls}-menu-submenu-title:hover`]: {
-            color: token?.layout?.sider?.colorTextMenuActive,
+            color: token.layout?.sider?.colorTextMenuActive,
             borderRadius: token.borderRadius,
-            backgroundColor: `${token?.layout?.header?.colorBgMenuItemHover} !important`,
+            backgroundColor: `${token.layout?.header?.colorBgMenuItemHover} !important`,
             [`${token.antCls}-menu-submenu-arrow`]: {
-              color: token?.layout?.sider?.colorTextMenuActive,
+              color: token.layout?.sider?.colorTextMenuActive,
             },
           },
         },
         [`${token.antCls}-menu-item-selected`]: {
-          color: token?.layout?.sider?.colorTextMenuSelected,
+          color: token.layout?.sider?.colorTextMenuSelected,
         },
         [`${token.antCls}-menu-submenu-selected`]: {
-          color: token?.layout?.sider?.colorTextMenuSelected,
+          color: token.layout?.sider?.colorTextMenuSelected,
         },
         [`&${token.antCls}-menu:not(${token.antCls}-menu-inline) ${token.antCls}-menu-submenu-open`]:
           {
-            color: token?.layout?.sider?.colorTextMenuSelected,
+            color: token.layout?.sider?.colorTextMenuSelected,
           },
 
         [`&${token.antCls}-menu-vertical`]: {
           [`${token.antCls}-menu-submenu-selected`]: {
             borderRadius: token.borderRadius,
-            color: token?.layout?.sider?.colorTextMenuSelected,
+            color: token.layout?.sider?.colorTextMenuSelected,
           },
         },
 
         [`${token.antCls}-menu-submenu:hover > ${token.antCls}-menu-submenu-title > ${token.antCls}-menu-submenu-arrow`]:
           {
-            color: token?.layout?.sider?.colorTextMenuActive,
+            color: token.layout?.sider?.colorTextMenuActive,
           },
 
         [`&${token.antCls}-menu-horizontal`]: {
@@ -122,20 +122,20 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
           ${token.antCls}-menu-submenu-active`]: {
             borderRadius: 4,
             transition: 'none',
-            color: token?.layout?.header?.colorTextMenuActive,
-            backgroundColor: `${token?.layout?.header?.colorBgMenuItemHover} !important`,
+            color: token.layout?.header?.colorTextMenuActive,
+            backgroundColor: `${token.layout?.header?.colorBgMenuItemHover} !important`,
           },
 
           [`${token.antCls}-menu-item-open,
           ${token.antCls}-menu-submenu-open,
           ${token.antCls}-menu-item-selected,
           ${token.antCls}-menu-submenu-selected`]: {
-            backgroundColor: token?.layout?.header?.colorBgMenuItemSelected,
+            backgroundColor: token.layout?.header?.colorBgMenuItemSelected,
             borderRadius: token.borderRadius,
             transition: 'none',
-            color: `${token?.layout?.header?.colorTextMenuSelected} !important`,
+            color: `${token.layout?.header?.colorTextMenuSelected} !important`,
             [`${token.antCls}-menu-submenu-arrow`]: {
-              color: `${token?.layout?.header?.colorTextMenuSelected} !important`,
+              color: `${token.layout?.header?.colorTextMenuSelected} !important`,
             },
           },
           [`> ${token.antCls}-menu-item, > ${token.antCls}-menu-submenu`]: {
@@ -151,7 +151,7 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
 
       [`${token.proComponentsCls}-top-nav-header-base-menu`]: {
         [`&${token.antCls}-menu`]: {
-          color: token?.layout?.header?.colorTextMenu,
+          color: token.layout?.header?.colorTextMenu,
           [`${token.antCls}-menu-item a`]: {
             color: 'inherit',
           },
@@ -161,19 +161,19 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
             ${token.antCls}-menu-item-active,
             ${token.antCls}-menu-submenu-active, 
             ${token.antCls}-menu-submenu-title:hover`]: {
-            color: token?.layout?.header?.colorTextMenuActive,
+            color: token.layout?.header?.colorTextMenuActive,
             borderRadius: token.borderRadius,
             transition: 'none',
-            backgroundColor: token?.layout?.header?.colorBgMenuItemSelected,
+            backgroundColor: token.layout?.header?.colorBgMenuItemSelected,
             [`${token.antCls}-menu-submenu-arrow`]: {
-              color: token?.layout?.header?.colorTextMenuActive,
+              color: token.layout?.header?.colorTextMenuActive,
             },
           },
 
           [`${token.antCls}-menu-item-selected`]: {
-            color: token?.layout?.header?.colorTextMenuSelected,
+            color: token.layout?.header?.colorTextMenuSelected,
             borderRadius: token.borderRadius,
-            backgroundColor: token?.layout?.header?.colorBgMenuItemSelected,
+            backgroundColor: token.layout?.header?.colorBgMenuItemSelected,
           },
         },
       },
@@ -194,24 +194,24 @@ const compatibleStyle: GenerateStyle<ProLayoutToken> = (token) => {
         },
       },
       [`${token.antCls}-menu-item-selected`]: {
-        color: token?.layout?.sider?.colorTextMenuSelected,
+        color: token.layout?.sider?.colorTextMenuSelected,
       },
       [`${token.antCls}-menu-submenu-selected`]: {
-        color: token?.layout?.sider?.colorTextMenuSelected,
+        color: token.layout?.sider?.colorTextMenuSelected,
       },
       [`${token.antCls}-menu:not(${token.antCls}-menu-horizontal)`]: {
         [`${token.antCls}-menu-item-selected`]: {
           backgroundColor: 'rgba(0, 0, 0, 0.04)',
           borderRadius: token.borderRadius,
-          color: token?.layout?.sider?.colorTextMenuSelected,
+          color: token.layout?.sider?.colorTextMenuSelected,
         },
         [`${token.antCls}-menu-item:hover, 
           ${token.antCls}-menu-item-active,
           ${token.antCls}-menu-submenu-title:hover`]: {
-          color: token?.layout?.sider?.colorTextMenuActive,
+          color: token.layout?.sider?.colorTextMenuActive,
           borderRadius: token.borderRadius,
           [`${token.antCls}-menu-submenu-arrow`]: {
-            color: token?.layout?.sider?.colorTextMenuActive,
+            color: token.layout?.sider?.colorTextMenuActive,
           },
         },
       },
@@ -235,12 +235,12 @@ const genProLayoutStyle: GenerateStyle<ProLayoutToken> = (token) => {
         flexDirection: 'column',
         width: '100%',
         backgroundColor:
-          token?.layout?.pageContainer?.colorBgPageContainer || 'transparent',
+          token.layout?.pageContainer?.colorBgPageContainer || 'transparent',
         position: 'relative',
         paddingBlock:
-          token?.layout?.pageContainer?.paddingBlockPageContainerContent,
+          token.layout?.pageContainer?.paddingBlockPageContainerContent,
         paddingInline:
-          token?.layout?.pageContainer?.paddingInlinePageContainerContent,
+          token.layout?.pageContainer?.paddingInlinePageContainerContent,
         '&-has-page-container': {
           padding: 0,
         },
@@ -262,13 +262,8 @@ const genProLayoutStyle: GenerateStyle<ProLayoutToken> = (token) => {
         zIndex: 0,
         height: '100%',
         width: '100%',
-        background: token?.layout?.bgLayout,
+        background: token.layout?.bgLayout,
       },
-    },
-    [`${token.antCls}-menu-submenu-popup`]: {
-      backgroundColor: 'rgba(255, 255, 255, 0.42)',
-      '-webkit-backdrop-filter': 'blur(8px)',
-      backdropFilter: 'blur(8px)',
     },
   };
 };