+++ packages/provider/src/index.tsx
@@ -9,12 +9,15 @@ import { SWRConfig, useSWRConfig } from 'swr';
 import type { IntlType } from './intl';
 import { findIntlKeyByAntdLocaleKey, intlMap, zhCNIntl } from './intl';
 
+import dayjs from 'dayjs';
 import type { DeepPartial, ProTokenType } from './typing/layoutToken';
 import { getLayoutDesignToken } from './typing/layoutToken';
 import type { ProAliasToken } from './useStyle';
 import { proTheme } from './useStyle';
 import { defaultToken, emptyTheme } from './useStyle/token';
 import { merge } from './utils/merge';
+
+import 'dayjs/locale/zh-cn';
 
 export * from './intl';
 export * from './useStyle';
@@ -147,7 +150,7 @@ export type ParamsType = Record<string, any>;
 export type ConfigContextPropsType = {
   intl?: IntlType;
   valueTypeMap?: Record<string, ProRenderFieldPropsType>;
-  token?: ProAliasToken;
+  token: ProAliasToken;
   hashId?: string;
   hashed?: boolean;
   dark?: boolean;
@@ -305,6 +308,10 @@ const ConfigProviderContainer: React.FC<{
       return nativeHashId;
     }
   }, [nativeHashId, proProvide.hashed, props.hashed]);
+
+  useEffect(() => {
+    dayjs.locale(locale?.locale || 'zh-cn');
+  }, [locale?.locale]);
 
   const configProviderDom = useMemo(() => {
     const themeConfig = {