+++ packages/table/src/utils/columnRender.tsx
@@ -54,8 +54,8 @@ export const renderColumnsTitle = (item: ProColumns<any>) => {
   );
 };
 
-/** 判断可不可编辑 */
-function isEditableCell<T>(
+/** 判断是否为不可编辑的单元格 */
+function isNotEditableCell<T>(
   text: any,
   rowData: T,
   index: number,
@@ -114,7 +114,8 @@ export function columnRender<T>({
 
   const renderTextStr = renderText(text, rowData, index, action as ActionType);
   const mode =
-    isEditable && !isEditableCell(text, rowData, index, columnProps?.editable)
+    isEditable &&
+    !isNotEditableCell(text, rowData, index, columnProps?.editable)
       ? 'edit'
       : 'read';