+++ tests/layout/pageContainer.test.tsx
@@ -307,7 +307,13 @@ describe('PageContainer', () => {
     const { container, rerender, unmount } = render(
       <ProLayout>
         <PageContainer>
-          <FooterToolbar>
+          <FooterToolbar
+            stylish={() => {
+              return {
+                height: '100%',
+              };
+            }}
+          >
             <button type="button" key="button">
               qixian
             </button>