+++ packages/card/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.7](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@2.5.6...@ant-design/pro-card@2.5.7) (2023-07-24)
+
+**Note:** Version bump only for package @ant-design/pro-card
+
 ## [2.5.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@2.5.5...@ant-design/pro-card@2.5.6) (2023-07-19)
 
 ### Bug Fixes